requirements.txt
fastapi==0.118.0
jose==1.0.0
passlib==1.7.4
pydantic==2.11.10
SQLAlchemy==2.0.43
psycopg2-binary==2.9.11
uvicorn==0.29.0
python-dotenv==1.0.1
argon2-cffi==25.1.0